requirements.txt
Django>=5.2,<6
djangorestframework>=3.15
django-cors-headers>=4.3
